+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.19.3](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.19.2...v0.19.3) (2026-09-15)
+
+### Bug Fixes
+
+* horizontal scrolling is missing in mail queue ([#1379](https://github.com/zextras/carbonio-admin-console-ui/issues/1379)) ([ca36423](https://github.com/zextras/carbonio-admin-console-ui/commit/ca364233022e23389962aa87a757c401fb9e0c3e))
+
+## [0.19.2](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.19.1...v0.19.2) (2026-09-15)
+
+### Other changes
+
+* enhance error handling and success notifications in account… ([#1377](https://github.com/zextras/carbonio-admin-console-ui/issues/1377)) ([8ae6ed6](https://github.com/zextras/carbonio-admin-console-ui/commit/8ae6ed65b8855f51dca61a616127ca4b24315a15))
+* update volume allocation handling in advanced mailstores co… ([#1378](https://github.com/zextras/carbonio-admin-console-ui/issues/1378)) ([86f84fe](https://github.com/zextras/carbonio-admin-console-ui/commit/86f84fe6abdeab3aa1bc983f31cb901eae526124))
+
 ## [0.19.1](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.19.0...v0.19.1) (2026-09-08)
 
 ### Bug Fixes